Revert "feat: convert to typescript"
