+++ lib/core/deploy/deploy.js
