+++ lib/core/deploy/index.js
@@ -1,9 +1,11 @@
-import { pathExists, readJson, ensureFile, writeFile } from 'fs-extra';
-import { join } from 'path';
-import YAML from 'yaml';
-import { validate } from '../info';
-import { getCapabilityModule } from '../catalog';
-import { applyFromFile } from '../resources';
+'use strict';
+
+const { pathExists, readJson, ensureFile, writeFile } = require('fs-extra');
+const { join } = require('path');
+const YAML = require('yaml').default;
+const { validate } = require('../../core/info');
+const { getCapabilityModule } = require('../../core/catalog');
+const { applyFromFile } = require('../../core/resources');
 
 // Returns the name of the deployment file in the given directory
 function deploymentFile(targetDir) {
@@ -69,8 +71,7 @@ function apply(capName, resources, targetDir, capability, props) {
     const module = getCapabilityModule(capability);
     const df = deploymentFile(targetDir);
     const rf = resourcesFile(targetDir);
-
-    return Promise.resolve(validate(module.info(), props))
+    return new Promise((resolve, reject) => resolve(validate(module.info(), props)))
         .then(() => module.apply(capName, resources, targetDir, props))
         .then(res => writeResources(rf, res))
         .then(() => readDeployment(df))
