+++ lib/core/maven/index.js
@@ -1,6 +1,7 @@
-// @ts-ignore
-import { spawn } from 'child-process-promise';
-import { join } from 'path';
+'use strict';
+
+const { spawn } = require('child-process-promise');
+const { join } =  require('path');
 
 const jar = join(__dirname, 'maven-model-helper.jar');
