+++ lib/core/utils/index.js
