--- tslint.json
@@ -1,44 +1,0 @@
-{
-  "extends": ["tslint:recommended", "tslint-react", "tslint-config-prettier"],
-  "rules": {
-    "max-classes-per-file": false,
-    // Custom
-    "trailing-comma": [false, {"multiline": "always", "singleline": "never"}],
-    "interface-name": [false, "always-prefix"],
-    "only-arrow-functions": false,
-    "member-ordering": false,
-    // General
-    "no-console": [true,
-      "time",
-      "timeEnd",
-      "trace"
-    ],
-    "max-line-length": [
-      true,
-      140
-    ],
-    "no-string-literal": false,
-    "no-use-before-declare": true,
-    "object-literal-sort-keys": false,
-    "ordered-imports": false,
-    "no-empty-interface": false,
-    "no-empty": false,
-    "quotemark": [
-      true,
-      "single",
-      "avoid-escape"
-    ],
-    "variable-name": [
-      true,
-      "allow-leading-underscore",
-      "allow-pascal-case",
-      "ban-keywords",
-      "check-format"
-    ]
-  },
-  "linterOptions": {
-    "include": [
-      "core/**/*.ts"
-    ]
-  }
-}
